detector/requirements.txt
torch
torchvision
numpy
scipy
Pillow

kafka-python-ng